Refactor tests to distinguish core plugin and base
