+++ npm/src/cocooned/cocooned.js
@@ -1,0 +1,8 @@
+import { Base } from './base'
+import { coreMixin } from './plugins/core'
+
+class Cocooned extends coreMixin(Base) {}
+
+export {
+  Cocooned
+}
